Fix the behavior of repeatable options determination
Since Python 3, the function `__iter__` has been added to the string
objects. Which means the wrapper can consider as repeatable a non
repeatable option. The issue occurs if the docstring has an option with
a default valeu and at least another repeatable option. See #7 for more
details.

Fix the issue by type checking the default value to check if it's a
string before checking if it's iterable. It doesn't respect the duck
typing principal, but I believe for this particular case, it won't
matter.

Fix #7